+++ NEWS.md
@@ -1,5 +1,8 @@
 # Sparklyr 0.5.0 (UNRELEASED)
 
+- Added `sdf_quantile()` as a means of computing (approximate) quantiles
+  for a column of a Spark DataFrame.
+
 - Added support for `n_distinct(...)`, based on call to Hive function
   `count(DISTINCT ...)`. (#220)